.circleci: Use CentOS 7.4-based Docker image
The Debian one has been something of a Jenga tower, with gcc segfaulting
depending on the precise order in which I install clang and certain Rust
tooling.

Try CentOS instead, in hopes it will be more reliable.

This also addresses the issues with rustfmt not working as installed in
the Docker image.