Fix benchmarks
Benchmarks were not previously covered by CI, and broke due to API
changes.

This updates the benchmarks to use the new API, and adds them to CI to
ensure they aren't broken by future API changes.

Ideally we'd just ensure the benchmarks compile as opposed to actually
running them in CI, but I can't figure out exactly how to do that with
criterion.