ECDSA (NIST P-256) with *ring* provider
Adds support for using ECDSA with the NIST P-256 (a.k.a. prime256v1,
secp256r1) elliptic curve.

This includes several changes to the ECDSA API to abstract across the
differences between ECDSA as used with secp256k1 and ECDSA as used with
P-256. As an example, by convention secp256k1 public keys are
compressed, whereas P-256 keys are not. Which format is preferred is now
an associated constant of the elliptic curve type.

Includes test vectors from NIST. Verification and public key generation
are tested, but not signing as ECDSA uses a random nonce. Testing this
should be possible but tricky, as the API Signatory exposes hides the
random number generation.