Factor ECDSA PublicKey into compressed/uncompressed curve points
Changes ECDSA `PublicKey` into an enum with `CompressedPublicKey` and
`UncompressedPublicKey` variants that each carry a tag and are the
appropriate length for their elliptic curves.

Adds a `from_untagged_point` method that takes a raw uncompressed public
point (as a type safe `GenericArray` of appropriate length).

Removes vesitages of "ASN.1" in regard to public keys: the tags are not
ASN.1, but rather SECG curve point magic numbers defined in
SEC 1: Elliptic Curve Cryptography (Version 2.0) section 2.3.4 (page 11)

<http://www.secg.org/sec1-v2.pdf>