Dual license base64.rs and hex.rs as MIT+Apache 2.0
These are derived from code that was originally licensed as MIT-only,
and was therefore explicitly marked as such and called out in the
README.md.

The original author @Sc00bz has authorized the derived Rust code to be
dual licensed as MIT+Apache 2.0 like the rest of the codebase, so this
commit updates the documentation and comments to reflect that.