Use the 'signature' crate
Migrates the core signing traits to the ones we've been developing in
the RustCrypto project:

https://github.com/RustCrypto/signatures/tree/master/signature-crate

This presently pins to HEAD on the above repo (i.e. `a94234e`), as in
the course of implementing this some problems came up in the upstream
crate. The goal is to land this PR so everything compiles against that
revision, and further refine the API before doing another release of
the `signature` crate.