Update to signature v0.2
Uses the newly released version of the `signature` crate to complete the
work that was started in #160.

This both uses a released version of the `signature` crate with a
compatible API, and implements `signatory-secp256k1` using the new
custom derive support for `Signer` and `Verifier`.

It also adds support for the `DigestSignature` feature of the upstream
crate which associates the default hash function to use for a particular
signature type as an associated type.