Remove toplevel 'signature' re-exports; add encoding::Error
This removes many of the types previously re-exported from the toplevel
of the `signatory` crate. This includes all of the types that come from
the `signature` crate.

Additionally this splits out a `signatory::encoding::Error` which is
unique from `signatory::signature::Error` and used to handle
encoding/decoding strings as well as reading/writing from files, in
which case it's nice to provide some additional debugging information
like the filename.